Refactor Clang-Tidy configuration and improve type handling in Random and String classes for better safety and consistency

+++ src/Engine/Core/Container/String.cpp
@@ -59,9 +59,9 @@ namespace ludus::core
         while (src < end)
         {
             wchar_t wc = L'\0';
-            const long converted = std::mbrtowc(&wc, src, end - src, &state);
+            const size_t converted = std::mbrtowc(&wc, src, static_cast<size_t>(end - src), &state);
             
-            if (converted == static_cast<long>(-1) || converted == static_cast<long>(-2))
+            if (converted == static_cast<size_t>(-1) || converted == static_cast<size_t>(-2))
             {
                 // Invalid sequence, skip byte
                 ++src;
